changing intervals name + light format to int

+++ PlantGrowthSystem_Backend/Helpers/Intervals.cs
@@ -12,7 +12,7 @@ namespace PlantGrowthSystem_Backend.Helpers
         public int max_Humidity { get; set; }
         public float min_Temperature { get; set; }
         public float max_Temperature { get; set; }
-        public float min_Light { get; set; }
-        public float max_Light { get; set; }
+        public int min_Light { get; set; }
+        public int max_Light { get; set; }
     }
 }
